+++ internal/testing/mocks/client.go
@@ -6,9 +6,9 @@ import (
 
 // APIClient is the interface that all API clients must implement
 type APIClient interface {
-	CreateDeployment(deployment api.Deployment, response *string) error
-	CreateService(service api.Service, response *string) error
-	CreateIngress(ingress api.Ingress) (*api.Ingress, error)
+	UpsertDeployment(deployment api.Deployment, response *string) error
+	UpsertService(service api.Service, response *string) error
+	UpsertIngress(ingress api.Ingress) (*api.Ingress, error)
 	CreateVolume(volume api.Volume) error
 	CreateEnvironment(env api.Environment) (*api.Environment, error)
 }
@@ -17,16 +17,16 @@ type APIClient interface {
 var _ APIClient = (*MockAPI)(nil)
 
 // Implement the interface methods for MockAPI
-func (m *MockAPI) CreateDeployment(deployment api.Deployment, response *string) error {
-	return m.CreateDeploymentFunc(deployment, response)
+func (m *MockAPI) UpsertDeployment(deployment api.Deployment, response *string) error {
+	return m.UpsertDeploymentFunc(deployment, response)
 }
 
-func (m *MockAPI) CreateService(service api.Service, response *string) error {
-	return m.CreateServiceFunc(service, response)
+func (m *MockAPI) UpsertService(service api.Service, response *string) error {
+	return m.UpsertServiceFunc(service, response)
 }
 
-func (m *MockAPI) CreateIngress(ingress api.Ingress) (*api.Ingress, error) {
-	return m.CreateIngressFunc(ingress)
+func (m *MockAPI) UpsertIngress(ingress api.Ingress) (*api.Ingress, error) {
+	return m.UpsertIngressFunc(ingress)
 }
 
 func (m *MockAPI) CreateVolume(volume api.Volume) error {